Fixed problem with radDivDE1e13False (there was dot instead of ->) Run populateUiInformation.php to update mandelboxVariable formula

+++ mandelbulber2/src/fractal_formulas.cpp
@@ -4109,9 +4109,9 @@ void MandelboxVariableIteration(CVector4 &z, const sFractal *fractal, sExtendedA
 					&& (fractal->transformCommon.stopIterations - fractal->transformCommon.startIterations
 							 != 0))
 			{
-				int iterationRange =
+				double iterationRange =
 					fractal->transformCommon.stopIterations - fractal->transformCommon.startIterations;
-				int currentIteration = (aux.i - fractal->transformCommon.startIterations);
+				double currentIteration = (aux.i - fractal->transformCommon.startIterations);
 				tempVC += fractal->transformCommon.offset000 * currentIteration / iterationRange;
 			}
